requirements.txt
mysql-connector-python
python-dotenv
gensim
quickumls
spacy<3
pymed
pandas
nltk
